后台的商品类型管理优化

+++ laigou-basic/basic-util/src/main/java/cn/itsource/laigou/util/PageList.java
@@ -9,14 +9,14 @@ import java.util.List;
  */
 public class PageList<T> {
 
-    private Integer total = 0;
+    private Long total = 0L;
     private List<T> rows = new ArrayList<>();
 
-    public Integer getTotal() {
+    public Long getTotal() {
         return total;
     }
 
-    public void setTotal(Integer total) {
+    public void setTotal(Long total) {
         this.total = total;
     }
 
@@ -31,7 +31,7 @@ public class PageList<T> {
     public PageList() {
     }
 
-    public PageList(Integer total, List<T> rows) {
+    public PageList(Long total, List<T> rows) {
         this.total = total;
         this.rows = rows;
     }
